Observer Pattern Logging
Used Observer pattern to implement logging messages in the game.
LogEntryBuffer and LogWriter are singleton classes, so are called intially in main and LogWriter is attached in LogEntryBuffer List of Observers in main.

whenever we need to log.
first take instance using getinstance of LogEntryBuffer and call writelog function of it by passing the string log message as parameter.

you can view the log file in logs folder of project.